Add initial AIE dialect to represent stream switches.
This dialect is intended as a basis for representing placement and
routing information for AIE arrays.  Firstly, it can represent logical
flows of data between cores and DMAs.  Secondly, it can represent
detailed connections between cores and switchboxes and the internals
of switchboxes.  The internal connections can be used to code generate
stream switch configuration and extract the original flows, providing
a verification methodology.