[aie] Allow inlining of DMA programs.
DMA programs can be lengthy, so allow them to be expressed with
FuncOp and CallOp.  Generating code requires inlining the function call,
so we need some hooks to enable that.  In particular, we lie a little
bit and say that AIE::MemOp is Callable (even though nothing can call it).