[AIE] Working --aie-standard-lowering pass
Previously we lowered directly to LLVM, but this is somewhat fragile
and doesn't interact well with the rest of the LLVM lowering process.
This patch introduces a new path that first lowers into the standard dialect
and relies on the regular LLVM lowering processor (using the 'bare ptr'
calling convention for simplicity) for the bulk of the transformation.