[AIE] enable linking with external files
It's now possible to add a "link_with" attribute to a core, specifying
an object file.  This object file will be included in the final link
through the linker script.

Tests are currently enabled using the cardano tools to build the linked
binary.  This means that we need a mechanism for cmake to find it.
For the time being, the path is hardcoded.