[aie] Enable intrinsic wrapper infrastructure
The representation of function intrinsics differs slightly between
the Chess compiler and LLVM.  This patch adds a framework for adapting
from the LLVM representation to the Chess representation by linking
with a library.  This enables MLIR to generate code in a single
fashion and to compile/link with either LLVM or Chess.