[aie] Add ability to query shim types
Not all shim tiles are created equal.  Basically there are three types:
1) PL tiles that support stream connections to the PL.
2) NOC tiles that support stream connections to the PL and a ShimDMA connection
to the NOC.  However, the PL connections pass through a ShimMux.
3) the NPI tile in column zero that supports no connections.

These differences are important because writing to configuration registers
that don't exist result in a Slave Bus Error, which by default results
in an unrecoverable error in Linux.