re-add me_basic from Vitis 2019.2
It turns out that newer versions of this file contain multiple
symbols in .text sections.  The _main_init symbol has to be placed
first in the linked executable (at address 0).  However while the
chess linker can place individual symbols, the gnu linker can only
place entire sections, which results in the wrong symbol coming first.

There are a few potential solutions to this:
1) add a new start symbol that jump to _main_init
2) refactor the original me_basic.c file into separate .o files.