Added initial set of benchmarks for AIE
The runtime lib is updated with the performance counter class that
allows the use of performance counters. The benchmarks and calibration
tests have been added to the test/ folder, along with a README that
discusses the individual benchmarks and explains the measurement tools.

Also, now link ARM code with -lm by default.