Update useLock
It appears that we've actually been using this incorrectly since
the very beginning.  The 4th operand actually determines whether
the lock operation should block waiting for the state of the lock
to be correct for the operation or not.  Usually this should be '1'
indicating a blocking operation.  To avoid confusion, this patch
updates the argument to be an optional enumerated attribute.  If
no attribute is specified, then the default is to perform a blocking
operation.  In addition, all of the existing tests are updated to
use the new default operand.