Use vector instead of AccType for float Mul + FMA
The AIEVec Mul and FMA ops use the AIEVec AccType to model the
accumulators of the integer path. These accumulators are only present
in the integer path. In order to move data between vector registers
and accumulators AIEVec UPS and SRS ops are inserted.

Prior to this commit both integer and floating point AIEVec Mul and FMAs
use the AIEVec AccType. However, as a result of the floating point
AIEVec Mul and FMA ops outputting AIEVec AccTypes, AIEVec UPS
and SRS ops are unnecessarily inserted (and later ignored).

This commit has floating point AIEVec Muls and FMAs output normal
vectors instead of AIEVec AccTypes, and in doing so clean up the
floating point path so that the unnecessary AIEVec UPS and SRS ops
are not inserted.