Verify Mul/FMA results' types against operands'
While the left and right operands' types are checked to see if they
match, there was no check to see if an integer result was produced
by floating point operands and vice versa. This becomes problematic
when the widths of the operands and result are checked, as it makes
little sense to compare the width of a float against that of an integer.
This commit adds this check to ensure that the result's type corresponds
to that of the operands.