Fix lups/ups decision logic for C++ translation (#122)
#113 removed the aievec.acc type from the AIEVec dialect
and changed the logic deciding between emitting a lups or ups
from determining it from the UPS's source operand to its result.
This caused improper decision making, as the result is what represents
the accumulator widths to determine whether to use lups or ups.

#113 also failed to remove aievec.acc from the AIEVecToCpp tests,
which were updated by #120. However, #120 also changed the
emitted lups calls to ups, and this commit changes them back to the
intended lups.