Added AIECreateLogicalLocks (#112)
Currently, we require the ID of all lock operations to be explicitly specified. 

This patch makes the lock ID optional and provides a pass to determine appropriate lock IDs, assuming that the tile of each lock is fixed.  This pass is incremental: only locks without an ID will be assigned and locks that have an assigned ID will keep the same lock ID, enabling a mix of automatic and manual determination of lock IDs.