[AIE][AIEVec] Structure aie-vectorize as a pipeline
Right now, the AIEVectorize pass is monolithic and makes poor use of
MLIR infrastructure. This is the first of a series of patches aimed
at modernizing the pass and making it easier to tweak and extend.

I've changed the name from aie-vectorize to aie-affine-vectorize
in anticipation of the next patch, which will embed the affine
supervectorizer into the pipeline.

In the future, AIEVectorize should also be renamed to something like
LowerVectorToAIEVector, which matches better the semantics of the
pass.

Also in the future, the pass should be split into a series of
conversion and analysis passes that compose better, so we can build
more vectorization options reusing the existing code.