[aiecc] Remove extra include options
In order to find the toolchain in the sysroot, we need to have
a 'target' that includes 'linux' and for the 'lib/gcc/$target/$version'
directory to have a corresponding 'include/gcc/$target/$version'.
In some of our sysroots, it seems that we find a lib/gcc, but it
doesn't have a corresponding include/gcc directory.  Instead
force using '/usr/lib,include/gcc'.

Fixes #148