[aiecc] Remove extra include options (#184)
In order to find the toolchain in the sysroot, we need to have
a 'target' that includes 'linux' and for the 'lib/gcc/$target/$version'
directory to have a corresponding 'include/gcc/$target/$version'.
In some of our sysroots, it seems that we find a lib/gcc, but it
doesn't have a corresponding include/gcc directory.  Instead
force using '/usr/lib,include/gcc'.

In addition, it seems that the code searching for libXAIE wasn't
as careful as it could be and could return a valid empty string
when libXAIE is not found.  So we now check more explicitly
whether libXAIE is found or not.