Refactor core-to-standard lowering
Previously, each core could only be exported in isolation, along with
the 'main' function that executed the code for the core.  As a
result, we ended up with a large number of separate compile jobs.
This patch refactors the lowering so that all cores can be exported
together as a single executable file, without a 'main' symbol.  Using
updated linker scripts we substitute in the correct main symbol.
As a result, we only need separate linker processes for each core.
This tends to be faster for designs where the compile jobs are all
small, since total compilation time is dominated by the clang process
start overhead.