[aiecc] Add --unified/--no-unified options
Depending on the runtime of AIE compiler jobs, it may be more or
less beneficial to try to parallelize compilation.  If compile jobs
are very short relative to the process startup time, then it makes
sense to reduce the cost of starting processes by reducing parallelism.
However, when processes are cheap, relative to compilation, then it
makes more sense to run lots of parallel compilations.
This patch enables aiecc to run in either mode.
--unified compiles the code for all cores in a single
compile job.  The code for each core is distinguished by a
separate toplevel function which can call shared code.  There is still
a separate linker job for each core with a separate linker script.
--no-unified compiles and links the code for each core independently.

--unified is the default, --no-unified provides the prior behavior.

Note that with aggressive function inlining, --unified can result in
large amounds of code in a single compilation.  In order to improve
compile time, we currently reduce the automatic inlining threshold
significantly, which may have an impact on performance.

Note that this likely won't work without the following patch to
llvm-link:
commit a2dc9f367c6be90589626a575f61ebea5409eee0
Author: Stephen Neuendorffer <stephen.neuendorffer@xilinx.com>
Date:   Mon Nov 21 18:03:30 2022 -0800

    [llvm-link] Fix options of llvm-link