Fix stack allocation in the local tile memory
Previously, we fixed the stack at the start of the tile memory
space (at address 0x20000).  However, this address is in the south
memory, which means that the stack was broken for cores in row 1.
Unfortunately, there's no obvious fixed address where the stack can
live, since the addresses for the east and west cores change.
This patch fixes the linker script generation to always use the
tile local memory (Which may be east or west depending on the row).
Note that this is consistent with the allocation that AssignMemBuffers
is reserving for the stack.
(All of this is a bit clumsy: in reality, the stack should be allocated
like everything else, preferably based off of static analysis that
bounds the size of the stack.  But we don't have a mechanism for that
yet.)

We also make sure to reserve memory spaces at the boundary so
random data can't be put there.