[aiecc] Set default location of Peano
When built, cmake can be configured using PEANO_INSTALL_DIR
to point to the peano installation.  Pass this to aiecc so that
we can set the path correctly.

Note that this is, to some extent, an inperfect solution, because
the installation no longer has a relative path, so the resulting
directory cannot be relocated.  The --peano command line option
can be used to change this later.