Introduce AIETargetModel (#389)
This patch is a significant refactoring of the code AIE dialect to extract
many (but not all) of the device-specific information into a single place.
The intention is that an AIE.device() operation will now contain an AIE
design and specify the device target of the design by name.

This adjusts many aspects of what is legal and how designs are lowered
to account for new devices, particular AIE2 and beyond.  The
AIE.device() operation implements an interface (AIEArray) that can be
used by future operations to describe things that are beyond a simple
naming scheme, such as relocatable regions, or custom/experimental
devices.  The AIEArray interface returns an AIETargetModel, which
encapsulates information about a particular device.  This enables a
relatively flexible mapping between MLIR ops and TargetModel objects,
in addition to allowing C++ modeling of device relationships, such
as the facts that AIE2 devices will share many common properties.