Update linker script generation with stack size control (#401)
CoreOp now supports a 'stackSize' attribute, enabling the control
of the stack allocation.

Along the way, fix a bug where memory areas were not properly being
reserved (either with gnu linker scripts or bcf files).  This allowed
static allocations to potentially overlap with allocated communication
buffers.  Note that in order to keep static allocations independent
on different cores, we allow them only to be mapped into the local
memory of a core.

Fixes #209