refactor runtime_lib compilation
This patch enables compiling the host runtime library for multiple
host architectures (e.g., x86_64, aarch64).  This is done by specifying
AIE_RUNTIME_TARGETS as a list of targets and ${target}_TOOLCHAIN_FILE
for each target.  It also creates a build directory with the same
structure as the install directory, so that tests can work effectively.

Along the way, we separate out the aie_runtime_lib directory to contain
libraries for different aie target architectures.  The intention here
is to be able to build a toolchain which can be used to target different
host/device combinations at runtime, particularly, ARM + VCK190 (e.g.
aarch64 host and AIE1 target) and X86 + VCK5000 (e.g. x86_64 host and
AIE2 target).

This patch also moves to using the local compilation of libxaiengine,
rather than explicitly relying on pulling it out of a sysroot when
cross-compiling.  This normalizes the aiecc.py code for handling
this library between host and cross-compiles.