aiecc.py can compile simulation
Previously, aiecc.py generated a simulation directory with a
Makefile.  Running simulation required running 'make'.  However
the Makefile was fixed and made some assumptions about the structure
of a design.

Rather than fix the Makefile, this patch modifies aiecc.py to
implement the compilation.  The intention here is that with
(more or less) the same command line, one can retarget a design for
simulation.
After simulation is compiled it can be run with:
  aie.mlir.prj/sim.sh

The default has been changed to not generate and compile the simulation.
This can be added to most scenarios using aiecc.py --aiesim, with some
limitations:
1) Linking with xbridge is required, since it generates a map
file used by the simulator
2) Cross-compilation is not supported, since host libraries are likely
to be incompatible.  In this case, you'll need to build the design
separately without the --target option.