[test_library] refactor memory allocation from test library
The current implementation of memory allocation has several issues:
1) it statically allocates a set of memory descriptors, which prevents dynamic memory allocation and an extra API call
2) It is #ifdef'd on __AIESIM__, in order to distinguish between the simulation allocator and the libxaie allocator.  This requires us to have multiple builds of the library and switch between them
3) The libxaie allocator is going away, so needs to be replaced anyway.

This patch refactors this code into a separate library, limiting the amount of target/simulation specific code and making it easier to manage automatically in aiecc.  This makes sense because we pretty
much always need some sort of allocator.  It will also make it easier to provide other allocator implementations (e.g. for PCIe on VCK5000, or under windows for Phoenix, or for XRT)
For target use, we largely reuse code from libxaie that is going away, which is based on using /dev/ion.