[aiecc] simplify multi-line output
The python 'rich.progress' package by default parses output and
neatly breaks it across lines in a way that doesn't interfere with
the progress bar.  Unfortunately, this inserts newlines in 'verbose'
commands, making it hard to copy and paste them.  This patch disables
the default behavior, which might have small effects on the display
when the progress bar collides with other output, but makes it much
easier to cut out commands and replay them using -v.