When not cross-compiling, don't recompile LibXAIE
Vitis ships with a compiled libxaiengine.so.  Using this library
seems necessary when using the aiesimulator.  Generally, it also
seems redundant to recompile if we already have a binary from Vitis.
On the other hand, if we're cross-compiling, we still need to
compile libxaiengine ourselves, so continue to compile it from source