[xaie target] Correct lowering of locks in DMAs
In AIE1 devices, it is common for a DMA BD to acquire and then release
the same lock.  This resulted in a latent error where the release lock
specified in the IR was ignored.  In AIE2 devices, this error becomes
significant because the acquire and release lock for a BD are
usually different.  This patch fixes the error.