Separate unit_tests/04_shared_memory
There were two tests in this directory using the same .cpp file, but
different mlir files.  Unfortunately, in the current flow, this means
that the result products of these tests collide.  As a quick solution,
separate these into different directories.

Along the way, fix a latent bug in the 'row' version.