Fix CDO backend aie-rt dependency
The CDO backend requires a patched version of aie-rt.  Eventually,
this should just be wrapped into the shipped version of libxaiengine.so,
but for the time being we need a specially compiled version of it.

This patch also removes the old way of handling this, which
relied on a special build script to inject the patch.
Along the way, also refactor the cmake around aie-rt so that things
are nicely encapsulated