[aiecc] Be explicit about peano path (#794)
Historically, we assumed that peano was primarily accessed from the
user's path, which causes some issues where peano binaries conflict
with other LLVMs on the user path.  This patch removes this use
of the path in favor of referencing peano using explicit paths.
These explicit paths are discovered when aiecc runs, relative to
the location of the python code.  This replaces the (bogus)
compile-time setting of peano_install_dir, which was completely wrong
when distributing a packaged build.