Add bootgen as third-party tree
We compile a bootgen static library and the bootgen executable.
The static library will enable removing external calls to the bootgen
executable.  Update aie2xclbin and aiecc.py to use this.

Note that bootgen has relatively simple build configuration, but is
relatively warning-rich.  We disable alot of safety checking to get
a clean build.