simplifiy interface for cdo generation.
The cdo_driver exposes the 'byte_ordering' type.  However, exposing
this to all of the users of mlir-aie is somewhat heavyweight and
exposes the cdo_driver interface in lots of places.  Instead, expose
a simple boolean instead.  Even this is probably overkill since in
almost all cases we'll generating a little endian image.