[aievec] Move trunc/ext to common lowering
Currently, trunc/ext ops get lowered in a separate pass with two simple
patterns. This is mostly because other patterns check their users for
these ops, which makes all the patterns involved dependent on each other
and their order of application.

This patch addresses this issue and moves these two last patterns
together with all the other vector-to-aievec patterns.