[CIR] Added allow_unregistered_dialects to aiecc.py
This avoids choking on some remaining CIR attributes like:
Unable to parse module assembly:
error: "-":1:101: #"cir"<"lang<cxx>"> : 'none' attribute created with unregistered dialect. If this is intended, please call allowUnregisteredDialects() on the MLIRContext, or use -allow-unregistered-dialect with the MLIR opt tool used

This is simpler than adding the CIR dialect itself to the Python infrastructure
which is quite complex.