[aievec] Add reordering of extsi -> broadcast ops
Since, in MLIR, mixed-type arithmetic operations such as:

```mlir
%result = arith.<op> %operandA, %operandB : <narrow type> to <wide type>
```

Are canonically represented as:

```mlir
%wideA = arith.ext<ty> %operandA : <narrow type> to <wide type>
%wideB = arith.ext<ty> %operandB : <narrow type> to <wide type>
%result = arith.<op> %wideA, %wideB : <wide type>
```

This change should improve the chances of matching mixed-type arithmetic
ops by getting the sign extension operations closer to the arithmetic,
enabling the `ext/ext -> arith` pattern match when a broadcast op ends
up between one of the `ext` ops and the `arith` op.

The rewrite pattern is provided as a template so it can be reused
whenever more of these "swappable" `ext -> op -> arith` cases show up.