refactor chess interface
To properly support other architectures, we need to call xchesscc with
--aiearch as a parameter.  This makes it more awkward to used the
chess-clang hack.  It was always a hack anyway, so remove it and instead
require users to pass "+Wclang,xir" when feeding LLVMIR to chess.

Lastly, clean up a small issue where we really should be using
chess-llvm-link to link together downgraded LLVMIR for chess, rather
than mixing and matching our LLVM tools with chess tools.