Change aie.device op from NoTerminator to SingleBlockImplicitTerminator
There is a block merging pattern used by the greedy pattern rewriter
which queries a values's parent block's terminator. If the value
is an aie.buffer in an aie.device, the getTerminator() call will
assert because aie.device has no terminator. Another fix is to add
a check in the upstream code.